feat: adding support in Argentina for USD, and minor updates for ARS layout in the withdrawal flow
